+++ inject/inject-thrift-client/src/main/scala/com/twitter/inject/thrift/AndThenService.scala
@@ -1,8 +1,11 @@
 package com.twitter.inject.thrift
 
-import com.twitter.finagle.{Filter, Service}
-import com.twitter.scrooge.{ThriftMethod, ThriftStruct}
+import com.twitter.finagle.Filter
+import com.twitter.finagle.Service
+import com.twitter.scrooge.ThriftMethod
+import com.twitter.scrooge.ThriftStruct
 
+@deprecated("No replacement.", "2022-03-03")
 trait AndThenService {
   def andThen[Req <: ThriftStruct, Rep](
     method: ThriftMethod,