+++ inject/inject-thrift-client/src/main/scala/com/twitter/inject/thrift/internal/DefaultAndThenServiceImpl.scala
@@ -1,9 +1,12 @@
 package com.twitter.inject.thrift.internal
 
-import com.twitter.finagle.{Filter, Service}
+import com.twitter.finagle.Filter
+import com.twitter.finagle.Service
 import com.twitter.inject.thrift.AndThenService
-import com.twitter.scrooge.{ThriftMethod, ThriftStruct}
+import com.twitter.scrooge.ThriftMethod
+import com.twitter.scrooge.ThriftStruct
 
+@deprecated("No replacement.", "2022-03-03")
 private[thrift] class DefaultAndThenServiceImpl extends AndThenService {
   def andThen[Req <: ThriftStruct, Rep](
     method: ThriftMethod,